Add seeds for places
This takes a corpus of the existing data, so we have
something to work with. Because we want a reliable
way to link votes and places, we’re keeping the
original IDs. I’ve also added a `S3_HOSTNAME`
env variable, so we have a reliable way of
generating the cached images that we have
saved in s3