Most and least scenic places
I’ve ported the original queries to Postgres, but kept the principle of
using raw SQL due to the complexity of the calculations.

I’ve added a last resort option in case we don’t have enough data that
fulfils the default requirements to make it into top and bottom 5
respectively.