Rename ‘show’ action to ‘vote’
Initially it didn’t look like individual places would ever be shown, so
we thought it would be ok if we used the `show` action for voting on a
place.

However, individual places are shown from the leaderboard page.

In preparation for implementing that feature, we’re giving `vote` its
own action.