Add a cookie banner script
This assumes that all cookie-using JS has a `data-cookies`
element set and a `type="text/plain”` attribute (that prevents
the JS from running).

The code then checks for the presence of a `hasConsent`
cookie and  shows a banner if it has not been set. When a user
hits `accept` the cookie is set.

If the cookie is set, all JS with a `data-cookies` attribute is
activated by copying the script, and setting a type of
`text/javascript`, which allows it to load.

This approach is lovingly nicked from https://github.com/nhsuk/cookie-consent