Place picking service
Finding a place to show the user is becoming a complex operation, so
we’re moving it into a service.

The place selection follows the following priorities:
- a place that the user has not already rated (within this session)
- a place that has between 1 and 3 ratings, so the user can compare it
with theirs (in the spirit of this being a game)

The conditions become progressively less strict until a suitable place
is found.

In the unlikely event[1] that the user has rated everything, a random
place is picked.

[1] Unlikely in production where there are hundreds of thousands of
places.