Turn the leaderboard into a service
The logic for the leaderboard is complex enough to deserve its own
service class, rather than being a part of the `Vote` model.

This choice also makes it easier to see at a glance the default options
and their fallbacks.

We’ve changed the options as follows:
- maximum rating for bottom 5 is 2 by default, with 3 as a fallback[1]

[1] Ratings tend to be quite stern, so even a score of 5 for a built-up
area is a pretty good score.