Add link to the Data downloads page to footer
We can’t easily add a new link to the existing trio in the main menu,
due to the lack of visual assets that would integrate seamlessly into
the existing design.

As a compromise, and in addition to the data download page being linked
from the FAQ page, we’re also adding the link to the footer.