Silently ignore duplicate votes
Under normal operating conditions the user should never be shown a place
they've already rated - this is not under their control, and there’s
nothing they can do to avoid being shown a place they’ve already rated,
in the unlikely event that this happened.

The duplicate vote is not saved, but we are letting the user compare
their intended rating with other users (in the spirit of the game).

This was the original approach as well.