Send just rated place id as query param
Persisting the ID of the just-rated place in a session cookie leads to
confusion if the user browses to other pages on the site and then comes
back to the voting page, or if they open a new browser tab.

We are now sending it as a query parameter.

This has the inconvenient of exposing the query parameter to the user.
However, any non-existing places will be silently ignored.