Store the last rated place ID in the session
Here we revert to the “old” method of storing the last rated
place ID in the session, but instead of just fetching it, we fetch
and delete it using `Hash#delete`(https://apidock.com/ruby/Hash/delete). This way, when we refresh the page, we’re not
shown the place we last voted on again.