Calculate and store leaderboard data asynchronously
Calculating the stats each time someone loads the Leaderboard page is
slow and expensive. The leaderboard doesn't change very frequently, so
we can calculate it once per day and then display that.

We have chosen to serialise the leaderboard into a JSON file, which we
then deserialise for display.

The calculation is done via a rake task that can be invoked directly, or
via a scheduled cron job.

Co-authored-by: Lorna <lorna@dxw.com>
Co-authored-by: CristinaRO <cristina@dxw.com>