Select a place not rated by user at random
Previously, we were prioritising places with a small number of votes, in
order to balance the amount of data collected across places.

However, this translated to a slow and costly query, as most places got
more and more votes.

Rather than setting a moving target by upping the limit, it was decided
to remove the condition entirely, and rely on randomness.

We are still prioritising places that the current visitor has not rated
already.