Details on the leaderboard are less able to overlap
On wider screens we had the situation where a single long word would push the boundary of the grid layout and look wrong.

The word "Finchampstead" would not be split across multiple lines. Given the limited width of the container there was no where else for it to go except overflow outside.

We played with the idea of using a little CSS to keep the current layout but add an elipsis to long words:

```
.place {
  overflow: hidden;
  text-overflow: ellipsis;
}
```

The idea that this commit describes is slightly more involved. Instead of putting the picture and details side by side in a narrow layout, we give both it's own row and give each more space to fill. This means the image previews are slightly bigger and the text has more room

The consequence is that the pages on desktop are now slightly longer than before.