Centre leaderboard numbers within their badge
`vertical-align: middle;` wasn't being applied. The inspector reported that it needed to be used together with `display: inline`, which when added doesn't seem to have any effect.

The fixed positioning of `left` and `top` seem to make the situation worse so this was removed. The effect is that the badges lean more right across the image however the number is more centrally placed.