Leaderboard badge numbers align
The image wasn't of even proportions, it had more empty space on the top and left whilst on the right it was ended abruptly.

This approach first changed the image to a circle background created with CSS. Once we know we have an even circle we could then see it was the font was also putting the number off centre. If we change to arial the numbers are evenly placed but of course the font is different.

An outline is added in an effect to replicate the previous pattern.