Use absolute location for default photos folder
The images can be loaded from the homepage (AKA the voting page), or
from the show page of an individual place (linked from the leaderboard).

Previously, the default location (used when `IMAGE_HOSTNAME` is not
defined) was given as a relative location folder. Use absolute location
instead.

Note: if the `IMAGE_HOSTNAME` is defined in a production configuration
as a relative location, this value will also need to be modified.