Leaderboard numbers have a more flowery background
Instead of using a simple dotted outline we go further to try and replicate the previous flower shape.

We could added these styles to the li, li::before and li::after, however this is where the number is added and so my efforts to try and add the background ended up rotating the number too.

On mobile:
- the new badge background disappears unless we add top:0 and left:0.
- the alignment gets off centered by the inclusion of `@import "bootstrap/scss/bootstrap";` and so we have to be explicit about setting the box-sizing element to keep the badge the same on desktop and mobile