Generate 404, 422, and 500 pages
Instead of hardcoding them which results in out of date assets very
quickly, we generate error pages from templates like every other page.

In this process we've also improved the content of the pages to be more
useful for users who get them.

Note that the 503 page is expected while the server is offline, meaning
none of the linked assets would be available. Ideally we'd generate the
page, inline all of the local assets, and convert all images into data
URLs, but that hasn't been implemented at this point.

This task needs to be run on the server as part of the deploy step, too.