Fix footer spacing
Previous to this change, there was a white space above the footer due to
the margin added to create space for the grass graphic pushing the
container above it away. By moving that top margin to a bottom padding
on the container, we avoid that problem.

We also rationalise some of the styling to minimise the differences
between mobile and desktop, meaning the grass now appears on mobile,
too.