Upgrade to Bundler 2.4
Upgrade from Bundler 2.3.13 to 2.4.21.

This is an attempt to fix the following error that occurs during CI testing. This is caused by the new gem dependency sass-embedded,
which is required by dartsass-sprockets, which is required by bootstrap 5.3.2.

Your bundle is locked to sass-embedded (1.66.1-x86_64-linux) from rubygems
repository https://rubygems.org/ or installed locally, but that version can no
longer be found in that source. That means the author of sass-embedded
(1.66.1-x86_64-linux) has removed it. You'll need to update your bundle to a
version other than sass-embedded (1.66.1-x86_64-linux) that hasn't been removed
in order to install.

This error occurs despite the fact that the gem sass-embedded (1.66.1-x86_64-linux-gnu) is explicitly named in Gemfile.lock and
is available on rubygems.org. Discussion of related errors online suggests that upgrading Bundler to at least 2.3.22 may help.

Tests run fine locally with and without this commit.